+++ docker/sparta-marathon-utils.sh
@@ -26,6 +26,43 @@ function initJavaOptions() {
  fi
 }
 
+function initPluginCrossdata() {
+
+ if [ -v CROSSDATA_SECURITY_MANAGER_ENABLED ] && [ CROSSDATA_SECURITY_MANAGER_ENABLED == "true" ]; then
+    INFO "[GOSEC-CROSSDATA-CONFIG] Choose version: OK"
+    PLUGIN_FILES=(`ls -d -1  /opt/sds/crossdata/plugins{*,.*} | grep -e "dyplon-crossdata-${GOSEC_CROSSDATA_VERSION//./\\.}.*\.jar"`)
+    INFO "[GOSEC-CROSSDATA-CONFIG] Version choosed: OK"
+
+    case "${#PLUGIN_FILES[*]}" in
+            0)
+                ERROR "[GOSEC-CROSSDATA-CONFIG]GoSec Version (${GOSEC_CROSSDATA_VERSION}) is not compatible with the current version of Crossdata"
+                exit 1
+                ;;
+            1)
+                DYPLON_PLUGIN="${PLUGIN_FILES[0]}"
+                if [ -f  ${DYPLON_PLUGIN} ]; then
+
+                    INFO "[GOSEC-CROSSDATA-CONFIG] Copying Dyplon GoSec plugin in classpath"
+                    cp --preserve ${DYPLON_PLUGIN} /opt/sds/sparta/repo/
+                    INFO "[GOSEC-CROSSDATA-CONFIG] Copied Dyplon GoSec plugin in classpath"
+
+                else
+                    ERROR "[GOSEC-CROSSDATA-CONFIG]GoSec Version (${GOSEC_VERSION}) is not compatible with the current version of Crossdata"
+                    exit 1
+                fi
+                ;;
+            *)
+                ERROR "[GOSEC-CROSSDATA-CONFIG]More than 1 available plugin for provided GoSec Version (${GOSEC_VERSION})"
+                for each_plugin in "${PLUGIN_FILES[@]}"
+                do
+                    ERROR "[GOSEC-CROSSDATA-CONFIG]Available Dyplon GoSec plugin version: [$each_plugin]";
+                done
+                exit 1
+                ;;
+        esac
+ fi
+}
+
 function initSparkDefaultsOptions() {
 
   if [ -v PORT_SPARKUI ] && [ ${#PORT_SPARKUI} != 0 ]; then